requirements.txt
Django==2.0.4
djangorestframework==3.8.0
python-social-auth[django]==0.3.6
